add authorization

+++ backend/routes/posts.js
@@ -38,7 +38,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + '/images/' + req.file.filename
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
     });
     post.save()
       .then(createdPost => {
@@ -65,12 +66,17 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
 
-    Post.update({ _id: req.params.id }, post)
+    Post.update({ _id: req.params.id, creator: req.userData.userId }, post)
       .then(result => {
-        res.status(200).json({ message: 'Post updated successully!' });
+        if (result.nModified > 0) {
+          res.status(200).json({ message: 'Post updated successully!' });
+        } else {
+          res.status(401).json({ message: 'Not authorized.' });
+        }
       });
 });
 
@@ -114,11 +120,14 @@ router.get(
     })
 });
 
-router.delete('/:id', (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+router.delete('/:id', checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: 'Post deleted!' });
+      if (result.n > 0) {
+        res.status(200).json({ message: 'Post deleted!' });
+      } else {
+        res.status(401).json({ message: 'Not authorized.' });
+      }
     });
 });
